requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-docx==1.1.0
streamlit==1.32.0
python-dotenv==1.0.1
google-genai>=1.10.0
anthropic>=0.49.0
lancedb==0.9.0
pyarrow==16.0.0
sentence-transformers==3.0.0
transformers>=4.47.0
huggingface-hub>=0.34.0
PyMuPDF==1.24.0
fastapi==0.115.0
python-multipart==0.0.22
uvicorn==0.30.0
playwright==1.53.0
httpx>=0.27.0
langgraph==1.0.10
